chore: cleanup, graphite graphql and main graphql
