--- dashboard/src/components/common/InviteNotification/index.ts
@@ -1,1 +1,0 @@
-export { default as InviteNotification } from './InviteNotification';
