+++ dashboard/src/features/orgs/projects/common/components/ApplicationPaused/ApplicationPaused.tsx
@@ -9,10 +9,7 @@ import { StagingMetadata } from '@/features/orgs/projects/common/components/Stag
 import { useIsCurrentUserOwner } from '@/features/orgs/projects/common/hooks/useIsCurrentUserOwner';
 import { useCurrentOrg } from '@/features/orgs/projects/hooks/useCurrentOrg';
 import { useProject } from '@/features/orgs/projects/hooks/useProject';
-import {
-  GetAllWorkspacesAndProjectsDocument,
-  useUnpauseApplicationMutation,
-} from '@/generated/graphql';
+import { useUnpauseApplicationMutation } from '@/generated/graphql';
 import { useState } from 'react';
 
 export default function ApplicationPaused() {
@@ -23,9 +20,7 @@ export default function ApplicationPaused() {
     useState(false);
 
   const [showDeletingModal, setShowDeletingModal] = useState(false);
-  useUnpauseApplicationMutation({
-    refetchQueries: [{ query: GetAllWorkspacesAndProjectsDocument }],
-  });
+  useUnpauseApplicationMutation();
 
   return (
     <>
