+++ dashboard/src/features/orgs/projects/common/hooks/useCheckProvisioning/useCheckProvisioning.ts
@@ -1,12 +1,9 @@
 import { useIsPlatform } from '@/features/orgs/projects/common/hooks/useIsPlatform';
 import { useProject } from '@/features/orgs/projects/hooks/useProject';
-import {
-  GetAllWorkspacesAndProjectsDocument,
-  useGetApplicationStateQuery,
-} from '@/generated/graphql';
+import { useGetApplicationStateQuery } from '@/generated/graphql';
 import { ApplicationStatus } from '@/types/application';
 import { discordAnnounce } from '@/utils/discordAnnounce';
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 type ApplicationStateMetadata = {
   state: ApplicationStatus;
@@ -25,19 +22,10 @@ export default function useCheckProvisioning() {
     useState<ApplicationStateMetadata>({ state: ApplicationStatus.Empty });
   const isPlatform = useIsPlatform();
 
-  const { data, startPolling, stopPolling, client } =
-    useGetApplicationStateQuery({
-      variables: { appId: project?.id },
-      skip: !isPlatform || !project?.id,
-    });
-
-  async function updateOwnCache() {
-    await client.refetchQueries({
-      include: [GetAllWorkspacesAndProjectsDocument],
-    });
-  }
-
-  const memoizedUpdateCache = useCallback(updateOwnCache, [client]);
+  const { data, startPolling, stopPolling } = useGetApplicationStateQuery({
+    variables: { appId: project?.id },
+    skip: !isPlatform || !project?.id,
+  });
 
   const currentApplicationId = project?.id;
 
@@ -70,10 +58,6 @@ export default function useCheckProvisioning() {
         createdAt: data.app.appStates[0].createdAt,
       });
       stopPolling();
-      // Will update the cache and update with the new application state
-      // which will trigger the correct application component
-      // under `src\components\applications\App.tsx`
-      memoizedUpdateCache();
       return;
     }
     if (data.app.appStates[0].stateId === ApplicationStatus.Errored) {
@@ -89,15 +73,8 @@ export default function useCheckProvisioning() {
         `Application ${currentApplicationId} errored after provisioning: ${data.app.appStates[0].message}`,
       );
       stopPolling();
-      memoizedUpdateCache();
     }
-  }, [
-    data,
-    stopPolling,
-    memoizedUpdateCache,
-    currentApplicationId,
-    currentApplicationState.state,
-  ]);
+  }, [data, stopPolling, currentApplicationId, currentApplicationState.state]);
 
   return currentApplicationState;
 }
