--- dashboard/src/features/orgs/projects/common/hooks/useNavigationVisible/index.ts
@@ -1,1 +1,0 @@
-export { default as useNavigationVisible } from './useNavigationVisible';
