--- dashboard/src/features/orgs/projects/common/hooks/useProjectRoutes/index.ts
@@ -1,2 +1,0 @@
-export * from './useProjectRoutes';
-export { default as useProjectRoutes } from './useProjectRoutes';
