+++ dashboard/src/features/orgs/projects/deployments/components/DeploymentListItem/DeploymentListItem.tsx
@@ -13,10 +13,7 @@ import { useCurrentOrg } from '@/features/orgs/projects/hooks/useCurrentOrg';
 import { useProject } from '@/features/orgs/projects/hooks/useProject';
 import { execPromiseWithErrorToast } from '@/features/orgs/utils/execPromiseWithErrorToast';
 import type { DeploymentRowFragment } from '@/utils/__generated__/graphql';
-import {
-  GetAllWorkspacesAndProjectsDocument,
-  useInsertDeploymentMutation,
-} from '@/utils/__generated__/graphql';
+import { useInsertDeploymentMutation } from '@/utils/__generated__/graphql';
 import { formatDistanceToNowStrict, parseISO } from 'date-fns';
 import type { MouseEvent } from 'react';
 import { twMerge } from 'tailwind-merge';
@@ -56,9 +53,7 @@ export default function DeploymentListItem({
       })
     : '';
 
-  const [insertDeployment, { loading }] = useInsertDeploymentMutation({
-    refetchQueries: [{ query: GetAllWorkspacesAndProjectsDocument }],
-  });
+  const [insertDeployment, { loading }] = useInsertDeploymentMutation({});
   const { commitMessage } = deployment;
 
   async function redeployDeployment(event: MouseEvent<HTMLButtonElement>) {
