--- dashboard/src/gql/app/getAppPlansAndGlobalPlans.graphql
@@ -1,39 +1,0 @@
-fragment getAppPlanAndGlobalPlansApp on apps {
-  id
-  subdomain
-  workspace {
-    id
-    paymentMethods {
-      id
-    }
-  }
-  legacyPlan {
-    id
-    name
-  }
-}
-
-fragment getAppPlanAndGlobalPlansPlan on plans {
-  id
-  name
-  isFree
-  price
-  featureMaxDbSize
-}
-
-query getAppPlanAndGlobalPlans(
-  $workspaceSlug: String!
-  $appSubdomain: String!
-) {
-  apps(
-    where: {
-      workspace: { slug: { _eq: $workspaceSlug } }
-      slug: { _eq: $appSubdomain }
-    }
-  ) {
-    ...getAppPlanAndGlobalPlansApp
-  }
-  plans {
-    ...getAppPlanAndGlobalPlansPlan
-  }
-}
