--- dashboard/src/gql/app/getApplicationPlan.graphql
@@ -1,14 +1,0 @@
-query getApplicationPlan($workspace: String!, $slug: String!) {
-  apps(
-    where: { workspace: { slug: { _eq: $workspace } }, slug: { _eq: $slug } }
-  ) {
-    id
-    subdomain
-    legacyPlan {
-      name
-      price
-      upatedAt
-      featureMaxDbSize
-    }
-  }
-}
