--- dashboard/src/gql/app/getWorkspaceAndProject.graphql
@@ -1,5 +1,0 @@
-query GetWorkspaceAndProject($workspaceSlug: String!, $projectSlug: String) {
-  workspaces(where: { slug: { _eq: $workspaceSlug } }) {
-    ...Workspace
-  }
-}
