--- dashboard/src/gql/app/insertApp.graphql
@@ -1,12 +1,0 @@
-mutation insertApplication($app: apps_insert_input!) {
-  insertApp(object: $app) {
-    id
-    name
-    slug
-    workspace {
-      id
-      name
-      slug
-    }
-  }
-}
