--- dashboard/src/gql/fragments/workspace.gql
@@ -1,18 +1,0 @@
-fragment Workspace on workspaces {
-  id
-  name
-  slug
-  creatorUserId
-  workspaceMembers {
-    id
-    user {
-      id
-      email
-      displayName
-    }
-    type
-  }
-  projects: apps(order_by: { name: asc }) {
-    ...Project
-  }
-}
