--- dashboard/src/gql/payment-methods/deletePaymentMethod.graphql
@@ -1,5 +1,0 @@
-mutation deletePaymentMethod($paymentMethodId: uuid!) {
-  deletePaymentMethod(id: $paymentMethodId) {
-    id
-  }
-}
