--- dashboard/src/gql/payment-methods/getPaymentMethods.graphql
@@ -1,27 +1,0 @@
-fragment getPaymentMethods on paymentMethods {
-  id
-  createdAt
-  cardBrand
-  cardLast4
-  cardExpMonth
-  cardExpYear
-  isDefault
-  workspace {
-    id
-    apps {
-      id
-      legacyPlan {
-        isFree
-      }
-    }
-  }
-}
-
-query getPaymentMethods($workspaceId: uuid!) {
-  paymentMethods(
-    where: { workspaceId: { _eq: $workspaceId } }
-    order_by: { createdAt: desc }
-  ) {
-    ...getPaymentMethods
-  }
-}
