--- dashboard/src/gql/workspace-member-invites/deleteWorkspaceMemberInvite.graphql
@@ -1,5 +1,0 @@
-mutation deleteWorkspaceMemberInvites($id: uuid!) {
-  deleteWorkspaceMemberInvites(where: { id: { _eq: $id } }) {
-    affected_rows
-  }
-}
