--- dashboard/src/gql/workspace-member-invites/getWorkspaceMemberInvitesToManage.graphql
@@ -1,14 +1,0 @@
-query getWorkspaceMemberInvitesToManage($userId: uuid!) {
-  workspaceMemberInvites(where: { userByEmail: { id: { _eq: $userId } } }) {
-    id
-    email
-    userByEmail {
-      id
-    }
-    workspace {
-      id
-      name
-      slug
-    }
-  }
-}
