--- dashboard/src/gql/workspace-member-invites/insertWorkspaceMemberInvite.graphql
@@ -1,7 +1,0 @@
-mutation insertWorkspaceMemberInvite(
-  $workspaceMemberInvite: workspaceMemberInvites_insert_input!
-) {
-  insertWorkspaceMemberInvite(object: $workspaceMemberInvite) {
-    id
-  }
-}
