--- dashboard/src/gql/workspace-member-invites/updateWorkspaceMemberInvite.graphql
@@ -1,11 +1,0 @@
-mutation updateWorkspaceMemberInvite(
-  $id: uuid!
-  $workspaceMemberInvite: workspaceMemberInvites_set_input!
-) {
-  updateWorkspaceMemberInvites(
-    _set: $workspaceMemberInvite
-    where: { id: { _eq: $id } }
-  ) {
-    affected_rows
-  }
-}
