--- dashboard/src/gql/workspace-members/deleteWorkspaceMember.graphql
@@ -1,5 +1,0 @@
-mutation deleteWorkspaceMember($id: uuid!) {
-  deleteWorkspaceMember(id: $id) {
-    id
-  }
-}
