--- dashboard/src/gql/workspace-members/updateWorkspaceMember.graphql
@@ -1,8 +1,0 @@
-mutation updateWorkspaceMember(
-  $id: uuid!
-  $workspaceMember: workspaceMembers_set_input!
-) {
-  updateWorkspaceMember(_set: $workspaceMember, pk_columns: { id: $id }) {
-    id
-  }
-}
