--- dashboard/src/gql/workspace/deleteWorkspace.gql
@@ -1,5 +1,0 @@
-mutation deleteWorkspace($id: uuid!) {
-  deleteWorkspace(id: $id) {
-    id
-  }
-}
