+++ dashboard/src/types/application.ts
@@ -7,7 +7,6 @@ import type {
   GetProjectQuery,
   PermissionVariableFragment,
   SecretFragment,
-  WorkspaceFragment,
 } from '@/utils/__generated__/graphql';
 
 /**
@@ -39,7 +38,6 @@ export type DesiredState =
 
 export type ApplicationState = AppStateHistoryFragment;
 export type Deployment = DeploymentRowFragment;
-export type Workspace = WorkspaceFragment;
 export type Organization = GetOrganizationQuery['organizations'][0];
 // export type Project = ProjectFragment;
 export type Project = GetProjectQuery['apps'][0];
